make eloqstore as brpc module (#55)
* make eloqstore as brpc module
* notify external processor when receive new kv request
* add async io listener to poll the completed io
---------

+++ task_manager.cpp
@@ -72,7 +72,7 @@ void TaskManager::FreeTask(KvTask *task)
     }
 }
 
-uint32_t TaskManager::NumActive() const
+size_t TaskManager::NumActive() const
 {
     return num_active_;
 }
